requirements.txt
discord
aiosqlite
python-dotenv